feat: custom hook 사용하는 것으로 변경
