+++ src/components/bookcase/Cards.tsx
@@ -1,18 +1,70 @@
+import { useEffect, useState } from "react";
 import styled from "styled-components";
 
 import { BookcaseInfo } from "../../pages/Bookcase";
+import { getData } from "../../utils/lib/api";
+import { Loading } from "../common";
 import { AddBookCard, BookCard } from ".";
 import Empty from "./cardSection/Empty";
 
 interface CardsProps {
-  bookcaseInfo: BookcaseInfo[];
-  reloadBookcase: (key: string) => void;
   isLogin: boolean;
+  navIndex: number;
 }
-export default function Cards(props: CardsProps) {
-  const { bookcaseInfo, reloadBookcase, isLogin } = props;
 
-  if (bookcaseInfo.length === 0) {
+const TOKEN = localStorage.getItem("booktez-token");
+const localToken = TOKEN ? TOKEN : "";
+
+const useGetBookcase = (key: string, navIndex: number) => {
+  const [bookcaseInfo, setBookcaseInfo] = useState<BookcaseInfo[]>([]);
+  const [isLoading, setIsLoading] = useState<boolean>(true);
+
+  const getBookcase = async () => {
+    try {
+      const {
+        data: {
+          data: { books },
+        },
+      } = await getData(key, localToken);
+
+      console.log("books", books);
+      setBookcaseInfo(books);
+    } catch (err) {
+      console.log("err", err);
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  useEffect(() => {
+    getBookcase();
+  }, [navIndex]);
+
+  return { bookcaseInfo, isLoading };
+};
+
+export default function Cards(props: CardsProps) {
+  const { isLogin, navIndex } = props;
+  let path: string;
+
+  switch (navIndex) {
+    case 1:
+      path = "/pre";
+      break;
+    case 2:
+      path = "/peri";
+      break;
+    case 3:
+      path = "/post";
+      break;
+    default:
+      path = "";
+  }
+  const { bookcaseInfo, isLoading } = useGetBookcase(`/book${path}`, navIndex);
+
+  if (isLoading) {
+    return <Loading />;
+  } else if (bookcaseInfo.length === 0) {
     return (
       <StDefaultSection>
         <Empty />
@@ -23,7 +75,7 @@ export default function Cards(props: CardsProps) {
       <StSection>
         <AddBookCard />
         {bookcaseInfo.map((bookcaseInfo: BookcaseInfo, idx: number) => (
-          <BookCard key={idx} bookcaseInfo={bookcaseInfo} reloadBookcase={reloadBookcase} isLogin={isLogin} />
+          <BookCard key={idx} bookcaseInfo={bookcaseInfo} isLogin={isLogin} />
         ))}
       </StSection>
     );
