+++ src/pages/Bookcase.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 
 import { Cards, Navigation } from "../components/bookcase";
-import { Loading, MainHeader } from "../components/common";
+import { MainHeader } from "../components/common";
 import { getData } from "../utils/lib/api";
 
 export interface BookcaseInfo {
@@ -15,78 +15,9 @@ export interface BookcaseInfo {
 const TOKEN = localStorage.getItem("booktez-token");
 const localToken = TOKEN ? TOKEN : "";
 
-export const useGetBookcase = (key: string) => {
-  const [bookcaseInfo, setBookcaseInfo] = useState<BookcaseInfo[]>([]);
-  const [isLoading, setIsLoading] = useState<boolean>(true);
-
-  const getBookcase = async () => {
-    try {
-      const {
-        data: {
-          data: { books },
-        },
-      } = await getData(key, localToken);
-
-      console.log("books", books);
-      books.forEach((book: BookcaseInfo) => {
-        setBookcaseInfo((currentBook) => [...currentBook, book]);
-      });
-    } catch (err) {
-      console.log("err", err);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    getBookcase();
-  }, []);
-
-  return { bookcaseInfo, isLoading, getBookcase };
-};
-
 export default function Bookcase() {
   const [isLogin, setIsLogin] = useState<boolean>(true);
   const [navIndex, setNavIndex] = useState<number>(0);
-
-  let path: string;
-
-  switch (navIndex) {
-    case 1:
-      path = "/pre";
-      break;
-    case 2:
-      path = "/peri";
-      break;
-    case 3:
-      path = "/post";
-      break;
-    default:
-      path = "";
-  }
-
-  const [bookcaseInfo, setBookcaseInfo] = useState<BookcaseInfo[]>([]);
-  const [isLoading, setIsLoading] = useState<boolean>(true);
-
-  const getBookcase = async (key: string) => {
-    try {
-      const {
-        data: {
-          data: { books },
-        },
-      } = await getData(key, localToken);
-
-      setBookcaseInfo(books);
-    } catch (err) {
-      console.log("err", err);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    getBookcase(`/book${path}`);
-  }, [navIndex]);
 
   const handleChangeNavIndex = (idx: number) => {
     setNavIndex(idx);
@@ -116,7 +47,7 @@ export default function Bookcase() {
     <>
       <MainHeader>서재</MainHeader>
       <Navigation navIndex={navIndex} onChangeNavIndex={handleChangeNavIndex} />
-      {isLoading ? <Loading /> : <Cards bookcaseInfo={bookcaseInfo} reloadBookcase={getBookcase} isLogin={isLogin} />}
+      <Cards navIndex={navIndex} isLogin={isLogin} />
     </>
   );
 }
